refactor(policy): toPolicyReviewInput 留在 PolicyForm/TypeForm.tsx，由 SimpleTypeForm 匯入
不另開檔案搬移，TypeForm.tsx 的 diff 只剩一個 export。
